requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.6
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
et-xmlfile==1.1.0
eventlet==0.33.3
greenlet==2.0.2
idna==3.4
kombu==5.2.4
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
